requirements.txt
msal==1.31.0
requests==2.32.3
python-dotenv==1.0.1

pdfplumber==0.11.4
pymupdf==1.24.10
